[0.9.5] Unneeded import emitted if type used in array position
When a type was used in the element type position of an array it was not being de-aliased, which led us to including it as an import.
The fix is to call resolveTypeReference rather than computeTypeReference because resolveTypeReference will unwrapp the aliased type.
This more closely matches what we do in resolveGenericTypeReference (where inside that method we will call resolveTypeReference)
Fixes bug #828068